# prompb: Generate client proto using Opaque API, for parsing only.