… go.sum
… model/textparse/protobufparse_test.go
… prompb/buf.yaml
… prompb/io/prometheus/client/metrics.pb.go
… prompb/io/prometheus/client/v1/custom_test.go
… prompb/io/prometheus/client/v1/metrics.pb.go
… scrape/scrape_test.go
… scripts/genproto.sh