Fixes https://github.com/prometheus/prometheus/issues/14668

Depends on https://github.com/prometheus/prometheus/pull/15720

Second attempt after https://github.com/prometheus/prometheus/pull/15726

* `prompb: Generate client proto using Opaque API, for parsing only.`. This also creates a `io/prometheus/client/v1` proto that supports Opaque API that will be by default at some point.