@bwplotka 2024-12-30
Just OpaqueAPI, before tricks (:

```
benchstat -filter="/fmt:PromProto" gogo=scrape-loop-v1.txt opaqueAPI=scrape-loop-v4nolazy.txt opaqueAPILazy=scrape-loop-v4.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                 │    gogo     │             opaqueAPI              │            opaqueAPILazy             │
                                 │   sec/op    │   sec/op     vs base               │    sec/op     vs base                │
ScrapeLoopAppend/fmt=PromProto-2   54.48µ ± 8%   74.42µ ± 5%  +36.61% (p=0.002 n=6)   119.36µ ± 3%  +119.11% (p=0.002 n=6)

                                 │     gogo     │              opaqueAPI              │           opaqueAPILazy            │
                                 │     B/op     │     B/op      vs base               │     B/op      vs base              │
ScrapeLoopAppend/fmt=PromProto-2   75.58Ki ± 1%   64.09Ki ± 0%  -15.21% (p=0.002 n=6)   81.55Ki ± 1%  +7.90% (p=0.002 n=6)

                                 │    gogo    │             opaqueAPI              │            opaqueAPILazy            │
                                 │ allocs/op  │  allocs/op   vs base               │  allocs/op   vs base                │
ScrapeLoopAppend/fmt=PromProto-2   810.0 ± 0%   1613.0 ± 0%  +99.14% (p=0.002 n=6)   2218.0 ± 0%  +173.83% (p=0.002 n=6)

```